requirements.txt
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.98.0
greenlet==3.2.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
openai==1.70.0
pydantic==1.10.9
python-dotenv==1.0.0
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.27.0
tqdm==4.67.1
typing_extensions==4.13.0
urllib3==2.3.0
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
